refactor: create builder class for converting rows to avoid duplication
